pyproject.toml
[project]
name = "pox-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache>=0.12.3",
    "aioconsole>=0.8.2",
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "aiomysql>=0.3.2",
    "aiosqlite>=0.21.0",
    "discord-py>=2.6.4",
    "distro>=1.9.0",
    "edge-tts>=7.2.3",
    "gtts>=2.5.4",
    "lmstudio>=1.5.0",
    "markovify>=0.9.4",
    "matplotlib>=3.10.7",
    "mcstatus>=12.0.6",
    "mido>=1.3.3",
    "moviepy>=1.0.3",
    "pillow>=12.0.0",
    "piper-tts>=1.3.0",
    "profanityfilter>=2.1.0",
    "prompt-toolkit>=3.0.52",
    "pymojang>=2.0.6",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "roblox>=2.0.0",
    "thefuzz>=0.22.1",
    "websockets>=15.0.1",
]
